float to double inversion

+++ src/diffoptics/transforms/LookAtTransform.py
@@ -27,13 +27,13 @@ class LookAtTransform(BaseTransform):
         self.transform = torch.tensor([[left[0], new_up[0], dir_[0], pos[0]],
                                        [left[1], new_up[1], dir_[1], pos[1]],
                                        [left[2], new_up[2], dir_[2], pos[2]],
-                                       [0., 0., 0., 1.]])
+                                       [0., 0., 0., 1.]], dtype=torch.double)
         self.inverse_transform = torch.inverse(self.transform)
 
         # Sanity check (making sure that the inversion went well)
-        assert torch.allclose(torch.matmul(self.transform, self.inverse_transform), torch.eye(4), rtol=1e-04, atol=1e-06
-                              )
-        x = torch.randn(4, 4)
+        assert torch.allclose(torch.matmul(self.transform, self.inverse_transform), torch.eye(4, dtype=torch.double),
+                              rtol=1e-04, atol=1e-06)
+        x = torch.randn((4, 4), dtype=torch.double)
         assert torch.allclose(torch.matmul(self.transform, torch.matmul(self.inverse_transform, x)), x, rtol=1e-04,
                               atol=1e-06)
         assert torch.allclose(torch.matmul(self.transform, torch.matmul(self.inverse_transform, x)), x, rtol=1e-04,
